Expose animation label attribute through ComposeAnimation
The animation label will be used to represent the animation as text in
Android Studio when applicable. The first direct example is naming the tabs
representing the animations in the animation inspector.

For transition animations (the only supported type so far), we first try to
use the label set explicitly in TransitionAnimation, but we also fall back
to the class name of the transition states as an alternative.

Test: PreviewAnimationClockTest#animationLabelIsSetExplicitlyOrImplicitly
Bug: 161709804
Relnote: N/A
Change-Id: I17fb74bd189bc1e2cb54dda52879d726050fd013