PointerInputFilter that spits out MotionEvents
This PointerInputFilter also provides hooks to
allow the client code to act like an Android View
in terms of handling pointer input.

This includes an analog to returning true or false
from onTouchEvent and an analog to calling
requestDisallowInterceptTouchEvent on a parent view.

Relnote: "Implemented PointerInteropModifier which provides
the ability to receive MotionEvents and interact with Compose
as if you are an implementation of an Android View."

Fixes: 158674168
Test: ./gradlew compose:ui:ui:connectedCheck

Change-Id: Ieb08cb00b2a70e3b4263640200cb2219b1fa728c