Set default locale to US
SimpleDateFormat will use the default locale value if not explicitly
set. If the default locale is Arabic, SimpleDateFormat#format will
return a string that does not conform to the given pattern.

In ExifInterface, if setDateTime is called on a device with an Arabic
locale, it will write a string value that will not follow the EXIF
standard "YYYY:MM:DD HH:MM:SS" format. Thus, we need to set the
locale to Locale.US at all times.

Bug: 160674042
Test: N/A
Change-Id: I351e3dd0b97ce3938ef52b900aae4a5dcf104e3e