Add LazyColumnForIndexed/LazyRowForIndexed
There are cases when it is useful to have both item and index, and similarly to the common pattern forEach/forEachIndexed we are adding an extra overload which provides the index in the itemContent callback. We are not just adding index right for LazyColumnFor to not make simple usages look worse:
LazyColumnFor(list) { _, item -> Text(item) }
instead of just
LazyColumnFor(list) { Text(it) }

Relnote: Added LazyColumnForIndexed/LazyRowForIndexed - versions of LazyColumnFor/LazyRowFor which provides both index and item in the itemCallback. It is useful when in addition to a current item you need to know a current index.
Test: new tests, new demo for manual testing
Change-Id: I65ff336f568c18bd875157e67ece1f8da6985d4c