Swipe to dismiss component
This introduces a SwipeToDismiss component that can be dismissed by
swiping left or right. A background can be specified, which will be
stacked behind the content and exposed when the content is swiped.
SwipeToDismiss is implemented using Modifier.swipeable and it comes
with its own DismissState class which inherits from SwipeableState.

Test: N/A
Relnote: "Add an experimental material component SwipeToDismiss."
Change-Id: I129e5ce937cf6eca5e4e12e6564db810da0cc902