Add minimal gradle.properties to playground buidls
Latest AGP reads certain properties at configuration time which prevents
us from setting them via the settings file.

This CL adds a common shared properties file to playground projects that
only includes a subset of properties set in the AndroidX
gradle.properties file.
As that file is not allowed to include anything that is not defined in
androidX, the itegrity is checked in ValidatePlaygroundGradleProperties
task which is now included in the buildOnServer task dependencies for
the main AndroidX build.

Test: cd room && ./gradlew bOS
Change-Id: Ieb5e8a646872119b348718eaef3ab4729b4c8bbb