heifwriter: wait for encoder stop to complete when closing
Currently the MediaCodec stop/release is run on a handler thread.
HeifWriter stop doesn't wait for encoder stop to finish. When
running cts, if the codec stop fails, it's counted against the
next test running.

Wait for the stop code to actually run before return from close(),
so that if the MediaCodec stop throws, it's counted as a HeifWriter
test failure.

bug: 162851305
bug: 161454256
test: manually change test to loop 100x to catch the codec stop
      failure, and verify in log that the failure happens before test
      finishes (so that it's counted correctly).
Change-Id: Iab63d85618e8938df66801335d30bffe2c95d6f8