Create and start the Camera.
This change provides a full implementation of the VirtualSessionState
object which is used to create the CameraCaptureSession.

At the most basic level VirtualSessionState is responsible for waiting
until it has a valid CameraDevice and a valid set of Surfaces available
to use in order to create the CaptureSession. Once the CaptureSession is
configured, a RequestProcessor instance is created and attached to the
GraphProcessor, which can then submit any currently queued requests.

This change also includes GraphState, which is responsible for opening
the camera and wiring it to the VirtualSessionState when start is called
and for detaching them when stop or close is called.

Test: Refactored several unit tests to handle new behaviors.
Test: Integration test app starts and displays a viewfinder.
Test: Existing unit tests continue to pass.

Change-Id: I5bdb7d298c5294e07f59742091e3eee9abb9519d