Few fixes for insert + delete/insert unit tests + add and build upon unsubmitted insert unit tests written by @yrezgui
Insert now expects one and only one entity in the params, and it doesn't have to be first.
ContentAccessMediaStore _id columns are now nullable, to allow insertion without specifying id.
Fix @ContentUpdate processing to consider when it's a suspend function and make sure it still detects that the return type is Int.

Test: gw contentaccess:contentaccess-compiler:test
Test: gw contentaccess:integration-tests:testapp:cAT
Change-Id: I54117fc59666dd55af3b28465414e27f4bdcb8ac