Fixed issue where VectorAssets would not update
Relnote: "Changed Vector classes to no longer
be data classes as the same object instance is
used for composition.

Added mutableStateOf flag for VectorPainter
to conditionally re-draw if contents of the
vector change.

Refactored VectorComponent instance to be
part of VectorPainter as it is re-used
across compositions.

Updated GroupComponent and PathComponent
to have their name fields be mutable"

Change-Id: Ieae454490e274d51a03433b6506319f692f7d981
Fixes: 162931275 151378808
Test: Added test to VectorTest