Port AppSearchSchema away from using protos.
Having AppSearchSchema populate a proto directly requires us to have the
protos and proto lib compiled in to every backend, which will add 200kb
to everyone even if they aren't using the local backend.

It also prevents us from sharing code with the platform, where protos in
the sdk are not allowed.

Test: AppSearchSchemaTest
Bug: 163453135
Change-Id: I40a3c73ab146c9a3c52f003979479b9b0fefbcbe