Add hasAnimations method to ComposeViewAdapter
Add a method to ComposeViewAdapter that can be called from Android Studio to
detect if the corresponding @Composable has animations that can be opened in
the Animation Inspector. That information can be used by Android Studio, for
example, to hide the Animation Inspector toolbar icon if there are no
animations that can be inspected.

Bug: 163469060
Test: Updated ComposeViewAdapterTest
Change-Id: I6d94fc2f61af9bd32da240acd6ac69d67b762fc7