Adds a sample and doc clarification for ordering in InteractionState
Often in design systems (such as Material) the most recent Interaction should be the one represented by the system, but unfortunately `Set` does not have a guarantee of ordering. LinkedHashSet does, but is also mutable, which is not a contract we want for InteractionState, so instead we document this behavior and guarantee that ordering is maintained.

Also adds InteractionStateTest to verify this behavior.

Bug: b/161522042
Bug: b/146345816
Bug: b/152525426
Test: InteractionStateTest
Change-Id: Iac4b4c7f733788785454fc29bfe56ca2c479b294