Account for late dispatching of injected input
When the first input event of a gesture was enqueued, it would fix its
eventTime to the current time. However, there is inevitably a delay
between enqueuing and dispatching of the event. This means the first x
events are going to be "late" and sent immediately without delay until
the eventTime of event x+1 is aligned with the current time.

Fix this by introducing "gestureLateness", which captures this delay and
offsets all injected events by this value to make sure that the start of
the gesture is aligned with the current time when it is dispatched.

Bug: 165658128
Test: ./gradlew ui:ui-test:cC
Change-Id: I66647f2e4978ae415b0cc868f94d38810837bb82