Make SlicePermissionActivity an AppCompatActivity.
If it's a regular activity, it doesn't work well in apps that use the
night mode functionality of appcompat. More specifically, if night mode
is enabled in the provider app, the first time you open the permission
dialog, it will be in day mode, the second time, it will be in night
mode. Waiting a few seconds in between, it'll be day mode again.

Test: Manually on an Android O emulator.

Change-Id: Ic7e8660d377aaccbb18cb2052103bc7c81a1eed5