Fix a leak when removing a detached fragment
When removing a detached fragment that was not present in the back
stack it would get added as a tag on the container view in anticipation
of a transition being run but since it's detached there will be no
transition and the fragment will remain as a tag until another
transition is performed.

With this change the fragment is only added as a tag if it has a view.

Test: FragmentTransitionTest
Bug: 166489383
Change-Id: I4f768a80c30cf14b43a246695ff99f792a9561f2