MediaRouter: Let provider release route controller
This CL lets MediaRouteProvider request a client
to release a route controller.

It can be used to support "Stop" in the output switcher.

In order to do that, we add a new message being used by
new client and new service.
For old clients, the service sends provider descriptor,
which contains a "disabled" route descriptor, which
eventually make the client unselect the route.

Bug: 163095048
Test: Run support v7 demos, cast to a sample provider,
tap "Stop" in the output switcher and confirm that onRouteUnselected is called.
The demos are built without MediaTransferReceiver not to use media router2.
CLINET_VERSION_3 and CLIENT_VERSION_4 are both tested as well.

Change-Id: Ieb2e9e529e90188466f6d91fe9d20ca3e9f665aa