Introduce alternative for global test methods.
Adds new test methods that are defined on ComposeTestRule or as
extensions. This is preparation for deprecation of the global ones.

I have also moved majority of ComposeTestRule to commonMain otherwise
finder extensions would need to go to jvmMain.

Bug: 167565814
Test: Migrated an existing test as example of the usage
Relnote: "Added onNode and other global methods on ComposeTestRule as
the current global ones are going to be deprecated."

Change-Id: Ieae36a4b67a3190759e7284a638f8b755c06c1ec