Ensure view focus is restored when animations are skipped
In the case where you inflate a fragment in xml, if that fragment
calls requestFocus() once its view is created, we save the focus, but
the activity of the fragment has not yet been CREATED. Once the activity
is created, we properly cancel all fragment animations, but we only
restore focus when we executePendingOperations so the focus is not
restored.

Instead of restoring based on animations, we should always restore any
focused views we may have saved.

Test: FragmentLifecycleTest#focusInflatedView
Bug: 167581921
Change-Id: I41bffd85e8061df863c1de7fd7bff266156fc128