Bugfix: in ImageCapture update crop rect if the buffer has been rotated in HAL.
The original crop rect is calculated based on camera sensor buffer. On some devices, the buffer is rotated before being passed to users, in which case the crop rect also needs additional transformations.

To correct the crop rect, We need to figure out what HAL has done to the buffer, and apply the same transformation to the crop rect.

Bug: 162836760
Test: manual test and ./gradlew bOS
Change-Id: I30e77cc5bd3ba57072d2bdf2c92de393f5425a23