Fix bug in validating file save destination for image capture
To validate the save destination of image capture when the outputOption contains a file, try opening a write stream to the file. If it succeeds, the file is writable and the image capture save location is valid, otherwise it's invalid.

This validation previously used File.canWrite(), this works correctly for files that are actually created (e.g. using File.createTemp()). When creating a file object with `new File(...)`, the method constantly returns false. This was causing the tests in the rotation test app (uiwidgets/rotations) to fail.

Bug: 167697553
Test: ImageSaveLocationValidatorTest.java + ImageCapture rotation tests
Change-Id: I8c5656fcc5da0c477203b3b5c8bb450be148b323