Avoid usages of AnimationSet for entering Animations
Adding an AnimationSet within an AnimationSet leads to
visual artifacts when used with entering Animations
due to how AnimationSet is implemented - something
that Fragments cannot control.

While we can't control developers using a <set> within
a <set> in their own Animation XML, our usage of
an additional AnimationSet meant that even a single
<set> in XML was enough to trigger the visual artifacts
and blinking.

By removing our AnimationSet, we lose the ability to
properly complete the special effect when the Animation
ends, but because that is the existing behavior of
Fragments, this is probably the best we can do while
1) maintaining the ability to use Animations
2) Avoiding visual artifacts

Test: existing tests pass, tested in sample apps
BUG: 163084315
Change-Id: Ie17af24c3920bb097cfc90fba6ca55e0cd7199ff