Ensure views are removed from the container before being destroyed
When the FragmentManager is destroyed, it is possible that we never
enqueue a remove operation, which means the fragment view is not removed
from the container before being destroyed.

We should always enqueue the remove operation and allow the
SpecialEffectController to cancel any pendingOperations when the
fragmentManager is being destroyed.

RelNote: "Fragment views are now always removed from the container
before being destroyed."
Test: added check to StrictViewFragment
Bug: 163084550

Change-Id: Id587699a6ae9dd696f568f8e0a5304156be1cd82