Removing unnecessary layout from Wrapper.kt
The only thing this Layout does right now is add a modifier that clears focus if the user clicks on a non-focusable part of the screen.
Move this modifier to FocusManager and remove this Layout from Wrapper.kt
Also make FocusManager internal since all it's usages are within the same module.

Bug: 159743502
Relnote: Made FocusManager private.
Test: ./gradlew  compose:ui:ui:connectedCheck -Pandroid.testInstrumentationRunnerArguments.class=androidx.compose.ui.focus.SetRootFocusTest
Change-Id: I7872fd59410800f15727c3220796af1d270d5a7e