Update to the DaoProcessor to fix compiler error on Dao with a generic super type with Kotlin "primitives".
When a Kotlin interface we generate has a method with primitives, it generates two methods for compatibility. This fix implements logic for delegating the duplicate methods to the concrete method. This is done by going through the list of unannotated methods encountered and finding the corresponding annotated method taking into account the method name, return type and method parameters.

Test: Baseline tests pass without throwing compiler errors in DaoPrimitiveTest.kt
Bug: 160258066
Relnote: Update to the DaoProcessor to fix compiler error on Dao with a generic super type with Kotlin "primitives".
Change-Id: Ice6bbd327c1a2693bafd371da0bfdfea8769b4c8