Fix issues caused by interrupting exit effects
When removing a fragment using an exit effect, then reversing the exit
operation (by using the back button) to an enter effect, while exiting
the fragment only ever makes it to AWAITING_EXIT_EFFECTS so the view is
not destroyed. This means that when the fragment attempts to reenter it
does not go through createView() which gets the view ready for effects.

For this case, we need to:

1. Ensure we endViewTransition() on the exiting view since the one we
call in onAnimationEnd() wouldn't be executed until after the enter
aniamtion ended.
2. We need to prepare the entering view to be animated in by setting
its alpha to 0, then same as we should if it had gone through createView

RelNote: "Exiting fragments that have their effects reversed now
correctly cancel and restart with the proper entering effect."
Bug: 167092035
Test: FragmentAnimationTest.popReplaceOperationWithAnimation

Change-Id: I622265e9b4ebb210657aac2a5b97dd47061bfd5c