Decoy generation for compose JS runtime
Based on the exploration and the fixes from KT-40980,
the best idea to fix the problem with transformed @Composable functions
was to generate duplicated functions which will be the target for all
transforms.

This CL does exactly this, hiding decoys behind a compilation flag for now.

Test: run
`./gradlew :compose:compiler:compiler-hosted:integration-tests:testDebugUnitTest -Pandroidx.compose.multiplatformEnabled=true`

Change-Id: If17327b11489eea8f96d7736b70ea3da08fe7895