Refactor SearchResults, SearchResult.Result, MatchInfo and AppSearchManager.
SearchResults contains behaviour methods and remains its own class. This
class will not be shared with framework as-is because framework has a
different async mechanism.

To facilitate reuse with different backends, the class becomes an interface.

SearchResult contains data only, and becomes its own class to facilitate
sharing with framework.

MatchInfo is only used in the context of SearchResult, so it becomes an
inner class.

AppSearchManager is removed and replaced with AppSearchSession.

Bug: 162450968
Test: AppSearchManagerTest, SearchSpecTest
Relnote: N/A
Change-Id: Iac0f1fd9330422dfb30a6b4b552970dc50732f37