Handle cases where kotlin.Unit is java void
When a method returns `Unit`, kotlin converts it to `void`.
That doesn't mean it can never be kotlin.Unit though as it
actually does happen for suspend functions with Unit return
type (the arg for Continuation is Unit, not Void).

This CL handles that with a special type as there is no other
way to represent `void` in KSP.

I've also fixed a case where the vararg parameters were not
automatically converted to array. Apperantly, it is only
necessary in KSP when the source file is in java.

Bug: 160322705
Test: existing tests + MethodSpecHelperTest
Change-Id: If5653c56d0d60bd71a21a12258ade16101cf1887